PC_Service/requirements.txt
# ESP32 CYD PC Service Requirements
pyserial>=3.5
psutil>=5.9.0
